+++ ci/RunBazelAction.sh
@@ -121,7 +121,6 @@ if [[ "${WITH_SPCS_IMAGE}" = true ]]; then
         "--action_env=BASE_BATCH_GPU_IMAGE_PATH=${BASE_BATCH_GPU_IMAGE_PATH}"
         "--action_env=MODEL_LOGGER_PATH=${MODEL_LOGGER_PATH}"
         "--action_env=PROXY_IMAGE_PATH=${PROXY_IMAGE_PATH}"
-        "--action_env=VLLM_IMAGE_PATH=${VLLM_IMAGE_PATH}"
     )
 fi